requirements.txt
# Core IR / ML
sentence-transformers==2.7.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.2.0
pandas>=2.0.0
tqdm>=4.66.0

# Optional lexical baseline
rank-bm25>=0.2.2

# CLI / utilities
typer>=0.9.0
rich>=13.7.0
